Adding git hook to repo
This runs automatically on each commit. It posts a locked thread
to /r/houstonheadlinebot where the title is the commit's subject and
the content is the commit's full message.

As this hook requires the creds, no one else will be able to use it
for its intended purposes. This is in the repo purely for reference.

This probably needs to be corrected to only check the log for master
but until I have an update that warrants a new branch, it doesn't
really matter as I'm doing the thing you shouldn't do and committing
directly to master.